Add shared test components for project questions
