+++ integration_tests/pages/components/errorSummaryComponent.ts
@@ -1,0 +1,5 @@
+export default class ErrorSummaryComponent {
+  shouldShowErrorSummary(field: string, errorMessage: string) {
+    cy.get(`[data-cy-error-${field}]`).should('contain', errorMessage)
+  }
+}
