+++ integration_tests/pages/courseCompletions/process/projectPage.ts
@@ -1,20 +1,15 @@
-import {
-  EteCourseCompletionEventDto,
-  ProjectOutcomeSummaryDto,
-  ProviderTeamSummaryDto,
-} from '../../../../server/@types/shared'
+import { EteCourseCompletionEventDto } from '../../../../server/@types/shared'
 import paths from '../../../../server/paths'
 import { pathWithQuery } from '../../../../server/utils/utils'
-import SelectInput from '../../components/selectComponent'
+import ProjectQuestionsComponent from '../../components/projectQuestionsComponent'
 import BaseCourseCompletionsPage from './baseCourseCompletionsPage'
 
 export default class ProjectPage extends BaseCourseCompletionsPage {
-  readonly teamInput = new SelectInput('team')
-
-  readonly projectInput = new SelectInput('project')
+  readonly form: ProjectQuestionsComponent
 
   constructor() {
     super('Match with a project')
+    this.form = new ProjectQuestionsComponent()
   }
 
   static visit(courseCompletion: EteCourseCompletionEventDto) {
@@ -22,22 +17,5 @@ export default class ProjectPage extends BaseCourseCompletionsPage {
       form: '12',
     })
     return this.visitAndCheck(path)
-  }
-
-  selectTeam(team: ProviderTeamSummaryDto) {
-    this.teamInput.select(team.code)
-    cy.get('button').contains('Select team').click()
-  }
-
-  selectProject(project: ProjectOutcomeSummaryDto) {
-    this.projectInput.select(project.projectCode)
-  }
-
-  shouldShowTeamError() {
-    this.shouldShowErrorSummary('team', 'Choose a team')
-  }
-
-  shouldShowProjectError() {
-    this.shouldShowErrorSummary('project', 'Choose a project')
   }
 }
